Update application identifiers and properties in project files
Added multiple <PropertyGroup> entries in MauiPlate.csproj to set ApplicationIdGuid and ApplicationDisplayVersion for various target frameworks and configurations (Debug and Release). Also added a <Title> element with the value "Plate".

Modified the <Identity> element in Package.appxmanifest to change the Name attribute from "maui-plate-placeholder" to "com.alexbob.mauiplate" to update the application identifier.